Front-end removes QuickReferences. (#143)
* Front-end removes QuickReferences. ResourcesRoot routes to Resources tag.

* Creates Errors.js for error and error checking.
