+++ src/js/redux/Store.js
@@ -50,11 +50,11 @@ export const storeWagtailPage = (wagtailPage) => {
         case "elearning_content.LessonPage":
             store.dispatch({ type: ADDED_LESSON_PAGE, lesson: wagtailPage });
             break;
-        case "elearning_content.QuickReferenceRoot":
-            // Ignore the QuickReferences' root.
+        case "elearning_content.ResourcesRoot":
+            // Ignore the ResourcesRoot' root.
             return;
-        case "elearning_content.QuickReferencePage":
-            // Ignore QuickReferences.
+        case "elearning_content.ResourceArticle":
+            // Ignore ResourceArticle.
             return;
         case "wagtailtrans.TranslatableSiteRootPage":
             // Ignore the i18nized site's root.
