+++ src/riot/App.riot.html
@@ -30,10 +30,10 @@
             ></LessonPage>
             <Settings if="{state.contentType === 'settings'}"></Settings>
             <Resources
-                if="{state.contentType === 'resources'}"
+                if="{state.contentType === 'resources' || state.contentType === 'ResourcesRoot'}"
             ></Resources>
             <ResourceArticle
-                if="{state.contentType === 'QuickReferencePage'}"
+                if="{state.contentType === 'ResourceArticle'}"
                 wagtailPage="{state.wagtailPage}"
             ></ResourceArticle>
             <Profile if="{state.contentType === 'profile'}"></Profile>
