+++ src/riot/Screens/Authentication.riot.html
@@ -32,7 +32,7 @@
     </div>
 
     <script>
-        import { alertIfRequestWasMadeOffline } from "js/ErrorChecks";
+        import { alertIfRequestWasMadeOffline } from "js/Errors";
         import { login } from "js/AuthenticationUtilities";
 
         export default {
