+++ src/riot/Screens/Settings.riot.html
@@ -51,7 +51,7 @@
         import {
             alertIfRequestWasMadeOffline,
             alertIfBrowserBlocksNotifications,
-        } from "js/ErrorChecks";
+        } from "js/Errors";
         import { getPlatform } from "js/PlatformDetection";
 
         export default {
